Skeleton for Mediator Class
