--- code/json_reader.py
@@ -1,44 +1,0 @@
-import csv
-import json
-
-
-def read_dataset_csv():
-    object_to_return = {}
-    with open('../data/apache_ATDx_input.csv', mode='r') as file:
-        csv_file = csv.DictReader(file)
-        for rule in csv_file:
-            object_to_return[rule['projectKey']] = dict(rule)
-    return object_to_return
-
-
-def read_rules_csv():
-    object_to_return = {}
-    with open('../data/ar_rules.csv', mode='r') as file:
-        csv_file = csv.DictReader(file)
-        for rule in csv_file:
-            object_to_return[rule['id']] = dict(rule)
-    return object_to_return
-
-
-def save_dict_as_json(file_path, data):
-    with open(file_path, 'w') as outfile:
-        json.dump(data, outfile, indent=4, default=str)
-
-
-def read_json(path):
-    f = open(path, )
-    dictionary = json.load(f)
-
-    return dictionary
-
-
-def from_csv_to_json():
-    ar_rules = read_rules_csv()
-    sua = read_dataset_csv()
-
-    out_file = open('../data/ar_rules.json','w+')
-    json.dump(ar_rules, out_file, indent=4, default=str)
-
-    out_file = open('../data/apache_ATDx_input.json','w+')
-    json.dump(sua, out_file, indent=4, default=str)
-
